+++ src/components/artist-question-screen/artist-question-screen.e2e.test.tsx
