+++ src/components/audio-player/audio-player.e2e.test.tsx
@@ -1,7 +1,7 @@
 import React from 'react';
 import Enzyme, {shallow} from 'enzyme';
 import Adapter from 'enzyme-adapter-react-16';
-import {AudioPlayer} from '../audio-player/audio-player';
+import {AudioPlayer} from './audio-player';
 
 Enzyme.configure({adapter: new Adapter()});
