+++ src/components/authorization-screen/authorization-screen.e2e.test.tsx
