+++ src/components/genre-question-screen/genre-question-screen.e2e.test.tsx
