+++ src/components/question-screen-header/question-screen-header.test.tsx
@@ -1,10 +1,11 @@
 import React from 'react';
 import renderer from 'react-test-renderer';
-import MistakeSigns from '../mistake-signs/mistake-signs';
+import QuestionScreenHeader from './question-screen-header';
 
 describe(`QuestionScreenHeader`, () => {
   it(`renders correctly`, () => {
-    const tree = renderer.create(<MistakeSigns
+    const tree = renderer.create(<QuestionScreenHeader
+      gameTime={3}
       mistakes={2}
     />).toJSON();
