+++ src/components/welcome-screen/welcome-screen.e2e.test.tsx
@@ -1,7 +1,7 @@
 import React from 'react';
 import Enzyme, {shallow} from 'enzyme';
 import Adapter from 'enzyme-adapter-react-16';
-import WelcomeScreen from '../welcome-screen/welcome-screen';
+import WelcomeScreen from './welcome-screen';
 
 Enzyme.configure({adapter: new Adapter()});
