+++ src/components/welcome-screen/welcome-screen.test.tsx
@@ -1,6 +1,6 @@
 import React from 'react';
 import renderer from 'react-test-renderer';
-import WelcomeScreen from '../welcome-screen/welcome-screen';
+import WelcomeScreen from './welcome-screen';
 
 describe(`WelcomeScreen`, () => {
   it(`renders correctly`, () => {
