+++ src/components/win-screen/win-screen.e2e.test.tsx
@@ -1,7 +1,7 @@
 import React from 'react';
 import Enzyme, {shallow} from 'enzyme';
 import Adapter from 'enzyme-adapter-react-16';
-import WinScreen from '../win-screen/win-screen';
+import WinScreen from './win-screen';
 
 Enzyme.configure({adapter: new Adapter()});
