+++ src/components/win-screen/win-screen.test.tsx
@@ -2,7 +2,7 @@ import React from 'react';
 import renderer from 'react-test-renderer';
 import {BrowserRouter} from 'react-router-dom';
 
-import WinScreen from '../win-screen/win-screen';
+import WinScreen from './win-screen';
 
 describe(`WelcomeScreen`, () => {
   it(`renders correctly`, () => {
